@@ -12,18 +12,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.0",
-    "@angular/cdk": "^6.2.0",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/core": "^6.1.0",
+    "@angular/animations": "8.0.0",
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "6.0.0",
+    "@angular/http": "6.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "@angular/fire": "^6.0.2",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/material": "^6.4.7",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
+    "@angular/material": "^8.2.3",
     "@ng-bootstrap/ng-bootstrap": "^4.1.1",
     "@syncfusion/ej2-angular-calendars": "^17.1.42",
     "babel-polyfill": "^6.26.0",
@@ -33,15 +33,15 @@
     "moment": "^2.24.0",
     "ng-recaptcha": "^4.3.0",
     "ngx-material-timepicker": "^3.3.1",
-    "rxjs": "~6.2.0",
+    "rxjs": "~6.5.5",
     "xlsx": "^0.16.6",
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1000.8",
-    "@angular/cli": "^10.0.8",
-    "@angular/compiler-cli": "^6.1.10",
-    "@angular/language-service": "^6.1.0",
+    "@angular-devkit/build-angular": "~0.800.2",
+    "@angular/cli": "^8.0.2",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "8.0.0",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
@@ -56,7 +56,7 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "^2.9.2"
+    "typescript": "3.4.5"
   },
   "engines": {
     "node": "8.9.4",
